% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponsiveGuides.R
\name{responsiveGuide}
\alias{responsiveGuide}
\title{Responsive Guides - An htmlwidget for displaying plot guides}
\usage{
responsiveGuide(
  variables,
  color_scale = NULL,
  manual_colors = NULL,
  ramp_levels = 100,
  title = NULL,
  direction = "column",
  justify = "center",
  focused = NULL,
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{variables}{A vector of values to be displayed. Either continuous or discrete.}

\item{color_scale}{A color scale to use for the guide. If NULL, manual_colors must be provided.}

\item{manual_colors}{A named vector of colors to use for the guide. If NULL, color_scale must be provided.}

\item{ramp_levels}{The number of colors breaks used to display a continuous color ramp.}

\item{title}{The title of the guide.}

\item{direction}{The orientation of the guide. Either "row" or "column".}

\item{justify}{The alignment of the guide. Either "left", "center", or "right".}

\item{focused}{The value to highlight in the guide, useful if used interactively.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}
}
\description{
Responsive Guides - An htmlwidget for displaying plot guides
}
